+++ packages/api-derive/src/democracy/types.ts
@@ -2,9 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Vec } from '@polkadot/types';
-import type { AccountId, Balance, BlockNumber, Hash, PropIndex, Proposal, ReferendumIndex, ReferendumInfoTo239, Vote } from '@polkadot/types/interfaces';
+import type { AccountId, Balance, BlockNumber, Call, Hash, PropIndex, ReferendumIndex, ReferendumInfoTo239, Vote } from '@polkadot/types/interfaces';
 import type { PalletDemocracyReferendumStatus, PalletDemocracyVoteThreshold } from '@polkadot/types/lookup';
 import type { BN } from '@polkadot/util';
+import type { HexString } from '@polkadot/util/types';
 
 export interface AtBlock {
   at: BlockNumber;
@@ -22,13 +23,15 @@ export interface DeriveDemocracyLock {
 
 export interface DeriveProposalImage extends AtBlock {
   balance: Balance;
-  proposal?: Proposal;
+  proposal?: Call;
+  proposalHash?: HexString;
+  proposalLen?: number;
   proposer: AccountId;
 }
 
 export interface DeriveDispatch extends AtBlock {
   index: ReferendumIndex;
-  imageHash: Hash;
+  imageHash: HexString;
   image?: DeriveProposalImage;
 }
 
@@ -43,14 +46,14 @@ export interface DeriveProposal {
 
 export interface DeriveProposalExternal {
   image?: DeriveProposalImage;
-  imageHash: Hash;
+  imageHash: HexString;
   threshold: PalletDemocracyVoteThreshold;
 }
 
 export interface DeriveReferendum {
   index: ReferendumIndex;
   image?: DeriveProposalImage;
-  imageHash: Hash;
+  imageHash: HexString;
   status: PalletDemocracyReferendumStatus | ReferendumInfoTo239;
 }