+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Cater for latest Substrate `DispatchError` (additional preimage statuses)
+- Adjust `api.derive.democracy.*` to return `imageHash` as `HexString` (as applicable, supports current-gen preimages)
+- Update to latest Substrate metadata
+
+
 ## 9.5.2 Oct 16, 2022
 
 Changes: